requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
anthropic==0.40.0
redis>=5.0.1
pytest==7.4.3
langchain==0.0.350
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
aiofiles==23.2.1
numpy>=1.26.0
scikit-learn>=1.3.2
scipy>=1.11.4
PyJWT==2.8.0
